+++ packages/react-native/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @blobatar/react-native
 
+## 2.7.0
+
 ## 2.6.0
 
 ## 2.5.0